Move descriptor decoding methods from Decoder to DeviceData.
